+++ src/web_server/web_server_albumart.c
@@ -13,7 +13,6 @@
 
 #include "../../dist/src/sds/sds.h"
 #include "../../dist/src/mongoose/mongoose.h"
-#include "../../dist/src/mongoose/mongoose_compat.h"
 #include "../../dist/src/frozen/frozen.h"
 #include "../sds_extras.h"
 #include "../api.h"
@@ -64,7 +63,7 @@ void send_albumart(struct mg_connection *nc, sds data, sds binary) {
 
 //returns true if an image is served
 //returns false if waiting for mpd_client to handle request
-bool handle_albumart(struct mg_connection *nc, struct mg_http_message *hm, t_mg_user_data *mg_user_data, t_config *config, int conn_id) {
+bool handle_albumart(struct mg_connection *nc, struct mg_http_message *hm, t_mg_user_data *mg_user_data, t_config *config, long long conn_id) {
     //decode uri
     sds uri_decoded = sdsurldecode(sdsempty(), hm->uri.ptr, (int)hm->uri.len, 0);
     if (sdslen(uri_decoded) == 0) {
